test(components): moved create-init-state to testing-helpers
Jira issue: VEGA-3140
